refactor: update context menu styles and improve payload extraction logic
