+++ packages/vscode-extension/src/network/utils/requestFormatUtils.ts
@@ -4,18 +4,6 @@ function stringify(obj: unknown): string {
   return JSON.stringify(obj, null, 2);
 }
 
-export function formatJSONBody(body: unknown): string {
-  if (typeof body !== "string") {
-    return "No response body";
-  }
-  try {
-    const parsed = JSON.parse(body);
-    return stringify(parsed);
-  } catch {
-    return body;
-  }
-}
-
 export function formatUrlParams(url: string): string {
   try {
     const urlObj = new URL(url);
@@ -29,32 +17,18 @@ export function formatUrlParams(url: string): string {
   }
 }
 
-/**
- * Formats response body for copying
- */
-export function formatResponseBodyForCopy(body: unknown): string {
-  if (!body) {
-    return "No response body available";
+export function formatJSONBody(body: unknown): string {
+  if (typeof body !== "string") {
+    return "No response body";
   }
-  
-  if (typeof body === "string") {
-    try {
-      // Try to parse and re-stringify for pretty formatting
-      const parsed = JSON.parse(body);
-      return JSON.stringify(parsed, null, 2);
-    } catch {
-      // Return as-is if not valid JSON
-      return body;
-    }
+  try {
+    const parsed = JSON.parse(body);
+    return stringify(parsed);
+  } catch {
+    return body;
   }
-  
-  // For non-string bodies, stringify them
-  return JSON.stringify(body, null, 2);
 }
 
-/**
- * Extracts and formats request data as JSON for copying
- */
 export function getRequestJson(log: NetworkLog): string {
   if (!log.request) {
     return JSON.stringify({ error: "No request data available" }, null, 2);
@@ -77,9 +51,6 @@ export function getRequestJson(log: NetworkLog): string {
   return JSON.stringify(requestData, null, 2);
 }
 
-/**
- * Extracts and formats response data as JSON for copying
- */
 export function getResponseJson(log: NetworkLog): string {
   if (!log.response) {
     return JSON.stringify({ error: "No response data available" }, null, 2);
@@ -96,4 +67,30 @@ export function getResponseJson(log: NetworkLog): string {
   }
 
   return JSON.stringify(responseData, null, 2);
+}
+
+export function getRequestPayload(log: NetworkLog): string {
+  if (!log.request) {
+    return "No request data available";
+  }
+
+  const { url, postData } = log.request;
+
+  const urlParams = formatUrlParams(url);
+  const hasUrlParams = urlParams !== "{}";
+
+  if (postData && postData !== "") {
+    const bodyData = formatJSONBody(postData);
+
+    if (hasUrlParams) {
+      return `URL Parameters:\n${urlParams}\n\nRequest Body:\n${bodyData}`;
+    }
+    return bodyData;
+  }
+
+  if (hasUrlParams) {
+    return urlParams;
+  }
+
+  return "No request body";
 }
